moved css outside of the search component
