+++ frontend/src/components/search/SearchBar.css
@@ -1,0 +1,16 @@
+.custom-scrollbar {
+scrollbar-color: rgba(255, 255, 255, 0.3) rgba(255, 255, 255, 0.1);
+}
+.custom-scrollbar::-webkit-scrollbar {
+width: 2rem;
+}
+.custom-scrollbar::-webkit-scrollbar-track {
+background: rgba(255, 255, 255, 0.1);
+border-radius: 2rem;
+}
+.custom-scrollbar::-webkit-scrollbar-thumb {
+background: rgba(255, 255, 255, 0.3);
+border-radius: 2rem;
+border: 0.5rem solid transparent;
+background-clip: padding-box;
+}
